app/requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iso8601==2.1.0
kombu==5.5.4
Kurigram==2.2.7
packaging==25.0
prompt_toolkit==3.0.51
pyaes==1.6.1
pypika-tortoise==0.6.1
Pyrogram==2.0.106
pyromod==3.1.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tortoise-orm==0.25.1
typing_extensions==4.14.1
tzdata==2025.2
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
